Use Decimal.js for scalars instead of Numbers.
I've abstracted out the arithmetic to refer to a Field object which
provides methods to do arithmetic operations, and convert to and from JS
numbers and strings.

There's a DecimalField, which uses Decimal.js to do high-precision
decimal arithmetic. Among other things, it makes it possible to work with
large-exponent numbers like Avogadro's number.

There isn't yet a facility to pick which field you use at runtime, or to
configure decimal.js.